\name{my1pack}
\alias{my1pack-package}
\docType{package}
\title{package}
\description{HW2 package}
\details{
Problem1:generate a plot of nromally distributed number from centain interval
Problem2:Medal comparison in Winter Olympics games
}
\author{Yue Wu <ywu67@jhu.edu>}
\keyword{package}
\examples{
problem1('bmw1Series',dnorm,c(-5,5),20,TRUE)
data(problem2)
}


